Introducing automatic change tracking + major refactoring
 * The constructor syntax has changed  (on both document access strategies)
 * New API: `getPageRoot()` (on both document access strategies)
 * Removed the obsolete, timestamp-based change tracking
 * Removed all traces of the change notification events. No longer needed.
 * Added change tracking using the mutation-summary library
 * Now we accept filters for ignoring changes on some parts, and also
   ignore those parts while scanning. (This only works correctly if they
   are at the end of the document)
 * Don't freak out on corpus change; just update all the data structures
   incrementally, and publish a `corpusChanged` event
 * Enhanced the debug methods for verifying the integrity of our
   mapping data. (`testNodeMapping()` and `testAllMappings()`)
 * Renamed some of the internal fields and methods to start with _.
 * Various small fixes and debug output

Moved to a truly async API, hiding all internal details of scanning

Until now, some methods on the API of dom-text-mapper needed to do some
internal housekeeping, before they could actually calculate and return
correct results. But some of this house-keeping can not be done in a synchronos
way over PDF documents, and so to preserve comatibility on the API of
dom-text-mapper and all it's descendants, we are now moving to a fully async
API.

This means that before calling any of the impacted methods (listed below),
you should first call the `ready()` method with a callback, and the d-t-m
is ready for your queries, it will call your callback with a `state` object,
and you can then execute your queries on the returned `state` object.

This impacts the following methods:
 * `getInfoForPath()`
 * `getInfoForNode()`
 * `getContentForPath()`
 * `getLengthForPath()`
 * `getDocLength()`
 * `getCorpus()`
 * `getContextForCharRange()`
 * `getMappingsForCharRange()`
 * `getMappingsForCharRanges()`
 * `getIndexForPos()`

It is no longer supported to call these methods directly on d-t-m, they are
only exposed on the `state` object, which is passed with the callback,
after `ready()` has finished the housekeeping. (Even so, when calling
these methods, we still do the usual housekeeping stuff, because it's very
easy to disrupt the integrity of the data, even by routine operations.)

Update PDF for change tracking + async API

 * Added pdf-text-mapper (previously from Annotator's PDF plugin),
   which is a descendant of PageTextMapperCore
 * Some common code from DomTextMapper has been moved into a new class,
   TextMapperCore, which is now the common ancestor for DomTextMapper
   and PageTextMapperCore.
 * PageTextMapperCore has been updated to follow the API changes
   brought in by the automatic DOM change tracking and the async API.
 * PdfTextMapper has been updated to follow the same changes.
 * Added a _testAllMappings() method, for compatibility with dom-text-mapper
 * Cosmetic changes all around the place

Debugging and testing improvements

 * When discarding a node because it's empty, mark it (and all descendents of it)
   as irrelevent
 * When discarding a node because it's content is not present in it's parentn's
   content (so it's probably invisible), mark it (and all descendents of it)
   as mystery.
 * Improve the internal consistency checking methods to give more information,
   and add verbosity control. (Aso fix some bugs.)
 * Improve the internal consistenct checking methods to avoid giving false alarms
   on irrelevent and mystery nodes.
 * Improvement on reporting differences in data