Redesign incremental scanning
Deeply changed how the mapping information is updated upon incremental
rescanning. The whole concept of `escalation` is now gone.

Furthermore, we don't do full rescans ever more; when we notice a corpus
change, we can calculate how that will effect other nodes, and we just
update the data.

Enhanced the debug methods for verifying the integrity of our
mapping data. (`testNodeMapping()` and `testAllMappings()`)

Also renamed some of the internal fields and methods to start with _.

Fixes #17.